Add GenerateResume.sh script, and files required to create my pdf resume.
tex_*.tex are required to take the mmd2tex output and make it pdf generate-able.
When GenerateResume.sh is run it creates the JohnRegnerResume.html using the standard
markdown generator. And the resume.pdf is created by using a mmd2tex then adding in header
and footer stuff, running pdflatex on that then cleaning up the latex mess.